+++ lib/frame/struct.h
@@ -7,8 +7,8 @@
 
 STRUCT(evm_frame)
 {
-    int pc;
-    int gas;
+    uint64_t pc;
+    uint64_t gas;
     evm_bytes_t code;
     evm_mem_t m;
     evm_stack_t s;
@@ -17,7 +17,7 @@ STRUCT(evm_frame)
 STRUCT(evm_frame_o)
 {
     evm_bytes_t returndata;
-    int gas;
+    uint64_t gas;
 };
 
 #endif
